+++ src/runCheck.ts
@@ -1,36 +1,56 @@
 import chalk from "chalk";
 
 import { logDiff } from "./differ";
-import { Checker, ICheck } from "./workflow/checker";
+import { Check } from "./workflow/check";
+import { Checker } from "./workflow/checker";
 import { Merger } from "./workflow/merger";
 import { WorkflowResource } from "./workflow/resource";
 import { Workflow } from "./workflow/workflow";
 
-function logCheck(check: ICheck, forceUpdate: boolean) {
+function logCheck(check: Check, forceUpdate: boolean) {
     if (check.action === "equal") return;
-    const icon = Checker.getCheckIcon(check);
-    const color = {
-        added: chalk.green,
-        updated: chalk.blue,
-        deleted: chalk.yellow
-    }[check.action];
     if (!check.force || forceUpdate) {
-        console.log(
-            color(
-                `  ${icon}  ${chalk.bold(check.item)} will be ${chalk.bold(
-                    check.action
-                )}`
-            )
-        );
+        console.log(check.color(`  ${check.checkMessage}`));
         return;
     }
-    console.log(
-        chalk.grey(
-            `  ${icon}  ${chalk.bold(check.item)} can be ${
-                check.action
-            }, use ${chalk.bold("--force")} flag to apply`
-        )
-    );
+    console.log(chalk.grey(`  ${check.noForceMessage}`));
+}
+
+export function runCheck(
+    localContent: string,
+    remoteContent: string,
+    forceUpdate: boolean,
+    showDiff: boolean
+): boolean {
+    const localWorkflow = Workflow.fromString(localContent);
+    const remoteWorkflow = Workflow.fromString(remoteContent);
+    const checker = new Checker(forceUpdate, localWorkflow);
+
+    const errors = checker.getErrors();
+    if (errors.length) {
+        errors.forEach(error => console.log(chalk.red(`  ✗  ${error}`)));
+        console.log(
+            chalk.red(`  ✗  has ${chalk.bold("errors")} that prevent update`)
+        );
+        return false;
+    }
+    const newWorkflow = new Merger(true).merge(localWorkflow, remoteWorkflow);
+    const checks = checker.getChecks(newWorkflow);
+    const applyChecks = checks.filter(check => check.isApplied(forceUpdate));
+
+    checks.forEach(check => logCheck(check, forceUpdate));
+    if (!applyChecks.length) {
+        console.log(chalk.grey("  ✓  is up to date"));
+        return true;
+    }
+    console.log(chalk.green("  ✓  can be updated"));
+    if (showDiff && remoteContent && localContent) {
+        const newContent = new Merger(forceUpdate)
+            .merge(localWorkflow, remoteWorkflow)
+            .render();
+        logDiff(localContent, newContent);
+    }
+    return true;
 }
 
 export async function runCheckAll(
@@ -42,57 +62,22 @@ export async function runCheckAll(
         items.map(item => item.getRemote())
     );
     const localContents = await Promise.all(items.map(item => item.getLocal()));
-    let noErrors = true;
-    items.forEach((item, index) => {
-        const remoteContent = remoteContents[index];
-        const localContent = localContents[index];
-        if (!remoteContent) {
-            console.log(chalk.red("  ✗  download failed"));
-            return;
-        }
-        if (!localContent) {
-            console.log(chalk.green("  ✓  will be created"));
-            return;
-        }
+    const statuses: Array<boolean> = items.map((item, index) => {
         const title = item.title || item.name;
         console.log(`${chalk.bold(title)} ${chalk.grey(item.path)}`);
 
-        const localWorkflow = Workflow.fromString(localContent);
-        const remoteWorkflow = Workflow.fromString(remoteContent);
-        const checker = new Checker(forceUpdate, localWorkflow);
+        const remoteContent = remoteContents[index];
+        const localContent = localContents[index];
+        if (!remoteContent) {
+            console.log(chalk.red(`  ✗  download failed: ${item.url}`));
+            return true;
+        }
+        if (!localContent) {
+            console.log(chalk.green("  ✓  will be created"));
+            return true;
+        }
 
-        const errors = checker.getErrors();
-        if (errors.length) {
-            errors.forEach(error => console.log(chalk.red(`  ✗  ${error}`)));
-            console.log(
-                chalk.red(
-                    `  ✗  has ${chalk.bold("errors")} that prevent update`
-                )
-            );
-            return;
-        }
-        const newWorkflow = new Merger(true).merge(
-            localWorkflow,
-            remoteWorkflow
-        );
-        const checks = checker
-            .getChecks(newWorkflow)
-            .filter(check => check.action !== "equal");
-        const applyChecks = checks.filter(({ force }) => !force || forceUpdate);
-
-        checks.forEach(check => logCheck(check, forceUpdate));
-        if (!applyChecks.length) {
-            console.log(chalk.grey("  ✓  is up to date"));
-            return;
-        }
-        console.log(chalk.green("  ✓  can be updated"));
-        if (showDiff && remoteContent && localContent) {
-            const newContent = new Merger(forceUpdate)
-                .merge(localWorkflow, remoteWorkflow)
-                .render();
-            logDiff(localContent, newContent);
-        }
-        noErrors = noErrors && errors.length === 0;
+        return runCheck(localContent, remoteContent, forceUpdate, showDiff);
     });
-    return noErrors;
+    return statuses.filter(x => !x).length === 0;
 }