+++ src/workflow/checker.ts
@@ -1,17 +1,7 @@
 import equal from "deep-equal";
 
+import { Check, TAction } from "./check";
 import { Workflow } from "./workflow";
-
-type TAction = "updated" | "deleted" | "added" | "equal";
-
-export interface ICheck {
-    highlight?: string;
-    action: TAction;
-    item: string;
-    force?: boolean;
-    checkMessage?: string;
-    updateMessage?: string;
-}
 
 export class Checker {
     force: boolean;
@@ -22,7 +12,7 @@ export class Checker {
         this.current = current;
     }
 
-    getChecks(update: Workflow): Array<ICheck> {
+    getChecks(update: Workflow): Array<Check> {
         return [
             ...this.getWorkflowChecks(update),
             ...this.getJobChecks(update),
@@ -57,104 +47,83 @@ export class Checker {
         force: boolean,
         oldValue: unknown,
         newValue: unknown
-    ): ICheck | null {
+    ): Check | null {
         const action = Checker.getAction(oldValue, newValue);
         if (!action) return null;
-        return {
-            item,
-            action,
-            force
-        };
+        return new Check(item, action, force);
     }
 
-    getWorkflowChecks(update: Workflow): Array<ICheck> {
+    getWorkflowChecks(update: Workflow): Array<Check> {
         return [
-            {
-                item: "top comment",
-                action: Checker.getAction(
+            new Check(
+                "top comment",
+                Checker.getAction(
                     this.current.commentLines,
                     update.commentLines
                 ),
-                force: true
-            },
-            {
-                item: "workflow name",
-                action: Checker.getAction(this.current.name, update.name),
-                force: true
-            },
-            {
-                item: "triggers",
-                action: Checker.getAction(
-                    this.current.triggers,
-                    update.triggers
-                ),
-                force: true
-            }
+                true
+            ),
+            new Check(
+                "workflow name",
+                Checker.getAction(this.current.name, update.name),
+                true
+            ),
+            new Check(
+                "triggers",
+                Checker.getAction(this.current.triggers, update.triggers),
+                true
+            )
         ];
     }
 
-    getJobChecks(update: Workflow): Array<ICheck> {
+    getJobChecks(update: Workflow): Array<Check> {
         const currentJob = this.current.job;
         const updateJob = update.job;
         return [
-            {
-                item: "environment",
-                action: Checker.getAction(currentJob.env, updateJob.env),
-                force: true
-            },
-            {
-                item: "runner",
-                action: Checker.getAction(currentJob.runsOn, updateJob.runsOn),
-                force: true
-            },
-            {
-                item: "strategy",
-                action: Checker.getAction(
-                    currentJob.strategy,
-                    updateJob.strategy
-                ),
-                force: true
-            },
-            {
-                item: "run condition",
-                action: Checker.getAction(currentJob.runsIf, updateJob.runsIf),
-                force: true
-            }
+            new Check(
+                "environment",
+                Checker.getAction(currentJob.env, updateJob.env),
+                true
+            ),
+            new Check(
+                "runner",
+                Checker.getAction(currentJob.runsOn, updateJob.runsOn),
+                true
+            ),
+            new Check(
+                "strategy",
+                Checker.getAction(currentJob.strategy, updateJob.strategy),
+                true
+            ),
+            new Check(
+                "run condition",
+                Checker.getAction(currentJob.runsIf, updateJob.runsIf),
+                true
+            )
         ];
     }
 
-    getStepChecks(update: Workflow): Array<ICheck> {
+    getStepChecks(update: Workflow): Array<Check> {
         const currentSteps = this.current.job.steps;
         const updateSteps = update.job.steps;
         return [
             ...currentSteps
                 .filter(step => step.findIndex(updateSteps) < 0)
-                .map(
-                    step =>
-                        ({
-                            item: `${step.name} step`,
-                            action: "deleted"
-                        } as ICheck)
-                ),
-            ...updateSteps.map(step => ({
-                item: `${step.name} step`,
-                action: (() => {
-                    const stepIndex = step.findIndex(currentSteps);
-                    if (stepIndex < 0) return "added";
-                    const localStep = currentSteps[stepIndex];
-                    if (step.isEqual(localStep)) return "equal";
-                    return "updated";
-                })() as TAction
-            }))
+                .map(step => new Check(`${step.name} step`, "deleted")),
+            ...updateSteps.map(
+                step =>
+                    new Check(
+                        `${step.name} step`,
+                        (() => {
+                            const stepIndex = step.findIndex(currentSteps);
+                            if (stepIndex < 0) return "added";
+                            const localStep = currentSteps[stepIndex];
+                            if (!step.isManaged()) return "kept";
+                            if (step.equals(localStep)) return "equal";
+                            return "updated";
+                        })() as TAction
+                    )
+            )
         ];
-    }
-
-    static getCheckIcon(check: ICheck): string {
-        return {
-            equal: "✓",
-            updated: "↻",
-            deleted: "✖",
-            added: "✎"
-        }[check.action];
     }
 }