+++ CHANGELOG.md
@@ -5,8 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `--path` CLI argument to specify local path to workflows
+
+### Changed
+- sync `jobs.*.strategy` from remote with `--force`
+- logs format
+
 ### Fixed
 - `--version` reports a correct version
+- `--diff` shows changes between local and updated version instead of remote
 
 ## [0.4.0]
 ### Added